+++ cmd/zetae2e/config/config.go
@@ -58,8 +58,11 @@ func ExportContractsFromRunner(r *runner.E2ERunner, conf config.Config) config.C
 	conf.Contracts.Solana.GatewayProgramID = config.DoubleQuotedString(r.GatewayProgram.String())
 	conf.Contracts.Solana.SPLAddr = config.DoubleQuotedString(r.SPLAddr.String())
 
-	conf.Contracts.Sui.GatewayPackageID = config.DoubleQuotedString(r.GatewayPackageID)
-	conf.Contracts.Sui.GatewayObjectID = config.DoubleQuotedString(r.GatewayObjectID)
+	if r.SuiGateway != nil {
+		conf.Contracts.Sui.GatewayPackageID = config.DoubleQuotedString(r.SuiGateway.PackageID())
+		conf.Contracts.Sui.GatewayObjectID = config.DoubleQuotedString(r.SuiGateway.ObjectID())
+	}
+
 	conf.Contracts.Sui.FungibleTokenCoinType = config.DoubleQuotedString(r.SuiTokenCoinType)
 	conf.Contracts.Sui.FungibleTokenTreasuryCap = config.DoubleQuotedString(r.SuiTokenTreasuryCap)