+++ zetaclient/chains/evm/observer/inbound.go
@@ -43,7 +43,7 @@ func (ob *Observer) ProcessInboundTrackers(ctx context.Context) error {
 
 	for _, tracker := range trackers {
 		// query tx and receipt
-		tx, _, err := ob.TransactionByHash(ctx, tracker.TxHash)
+		tx, _, err := ob.transactionByHash(ctx, tracker.TxHash)
 		if err != nil {
 			return errors.Wrapf(
 				err,
@@ -77,11 +77,11 @@ func (ob *Observer) ProcessInboundTrackers(ctx context.Context) error {
 		// try processing the tracker for v1 inbound
 		switch tracker.CoinType {
 		case coin.CoinType_Zeta:
-			_, err = ob.CheckAndVoteInboundTokenZeta(ctx, tx, receipt, true)
+			_, err = ob.checkAndVoteInboundTokenZeta(ctx, tx, receipt, true)
 		case coin.CoinType_ERC20:
-			_, err = ob.CheckAndVoteInboundTokenERC20(ctx, tx, receipt, true)
+			_, err = ob.checkAndVoteInboundTokenERC20(ctx, tx, receipt, true)
 		case coin.CoinType_Gas:
-			_, err = ob.CheckAndVoteInboundTokenGas(ctx, tx, receipt, true)
+			_, err = ob.checkAndVoteInboundTokenGas(ctx, tx, receipt, true)
 		default:
 			return fmt.Errorf(
 				"unknown coin type %s for inbound %s chain %d",
@@ -141,14 +141,14 @@ func (ob *Observer) observeInboundInBlockRange(ctx context.Context, startBlock, 
 		Uint64("start_block", startBlock).Uint64("to_block", toBlock).Logger()
 
 	// task 3: query the incoming tx to TSS address (read at most 100 blocks in one go)
-	lastScannedTssRecvd, err := ob.ObserveTSSReceive(ctx, startBlock, toBlock)
+	lastScannedTssRecvd, err := ob.observeTSSReceive(ctx, startBlock, toBlock)
 	if err != nil {
 		logger.Error().Err(err).Msg("error observing TSS received gas asset")
 	}
 
 	// task 4: filter the outbounds from TSS address to supplement outbound trackers
 	// TODO: make this a separate go routine in outbound.go after switching to smart contract V2
-	ob.FilterTSSOutbound(ctx, startBlock, toBlock)
+	ob.filterTSSOutbound(ctx, startBlock, toBlock)
 
 	var (
 		lastScannedZetaSent              = startBlock - 1
@@ -203,17 +203,17 @@ func (ob *Observer) observeInboundInBlockRange(ctx context.Context, startBlock, 
 }
 
 func (ob *Observer) fetchLogs(ctx context.Context, startBlock, toBlock uint64) ([]ethtypes.Log, error) {
-	gatewayAddr, _, err := ob.GetGatewayContract()
+	gatewayAddr, _, err := ob.getGatewayContract()
 	if err != nil {
 		return nil, errors.Wrap(err, "can't get gateway contract")
 	}
 
-	erc20Addr, _, err := ob.GetERC20CustodyContract()
+	erc20Addr, _, err := ob.getERC20CustodyContract()
 	if err != nil {
 		return nil, errors.Wrap(err, "can't get erc20 custody contract")
 	}
 
-	connectorAddr, _, err := ob.GetConnectorContract()
+	connectorAddr, _, err := ob.getConnectorContract()
 	if err != nil {
 		return nil, errors.Wrap(err, "can't get connector contract")
 	}
@@ -248,7 +248,7 @@ func (ob *Observer) observeZetaSent(
 	}
 
 	// filter ZetaSent logs
-	addrConnector, connector, err := ob.GetConnectorContract()
+	addrConnector, connector, err := ob.getConnectorContract()
 	if err != nil {
 		// we have to re-scan from this block next time
 		return startBlock - 1, errors.Wrap(err, "error getting connector contract")
@@ -301,7 +301,7 @@ func (ob *Observer) observeZetaSent(
 		}
 		guard[event.Raw.TxHash.Hex()] = true
 
-		msg := ob.BuildInboundVoteMsgForZetaSentEvent(app, event)
+		msg := ob.buildInboundVoteMsgForZetaSentEvent(app, event)
 		if msg == nil {
 			continue
 		}
@@ -325,7 +325,7 @@ func (ob *Observer) observeERC20Deposited(
 	logs []ethtypes.Log,
 ) (uint64, error) {
 	// filter ERC20CustodyDeposited logs
-	addrCustody, erc20custodyContract, err := ob.GetERC20CustodyContract()
+	addrCustody, erc20custodyContract, err := ob.getERC20CustodyContract()
 	if err != nil {
 		// we have to re-scan from this block next time
 		return startBlock - 1, errors.Wrap(err, "error getting ERC20Custody contract")
@@ -370,7 +370,7 @@ func (ob *Observer) observeERC20Deposited(
 		if event.Raw.BlockNumber > beingScanned {
 			beingScanned = event.Raw.BlockNumber
 		}
-		tx, _, err := ob.TransactionByHash(ctx, event.Raw.TxHash.Hex())
+		tx, _, err := ob.transactionByHash(ctx, event.Raw.TxHash.Hex())
 		if err != nil {
 			// we have to re-scan from this block next time
 			return beingScanned - 1, errors.Wrapf(err, "error getting transaction %s", event.Raw.TxHash.Hex())
@@ -385,7 +385,7 @@ func (ob *Observer) observeERC20Deposited(
 		}
 		guard[event.Raw.TxHash.Hex()] = true
 
-		msg := ob.BuildInboundVoteMsgForDepositedEvent(event, sender)
+		msg := ob.buildInboundVoteMsgForDepositedEvent(event, sender)
 		if msg != nil {
 			_, err = ob.PostVoteInbound(ctx, msg, zetacore.PostVoteInboundExecutionGasLimit)
 			if err != nil {
@@ -398,13 +398,13 @@ func (ob *Observer) observeERC20Deposited(
 	return toBlock, nil
 }
 
-// ObserveTSSReceive queries the incoming gas asset to TSS address and posts to zetacore
+// observeTSSReceive queries the incoming gas asset to TSS address and posts to zetacore
 // returns the last block successfully scanned
-func (ob *Observer) ObserveTSSReceive(ctx context.Context, startBlock, toBlock uint64) (uint64, error) {
+func (ob *Observer) observeTSSReceive(ctx context.Context, startBlock, toBlock uint64) (uint64, error) {
 	// query incoming gas asset
 	for bn := startBlock; bn <= toBlock; bn++ {
 		// observe TSS received gas token in block 'bn'
-		err := ob.ObserveTSSReceiveInBlock(ctx, bn)
+		err := ob.observeTSSReceiveInBlock(ctx, bn)
 		if err != nil {
 			// we have to re-scan from this block next time
 			return bn - 1, errors.Wrapf(err, "error observing TSS received gas asset in block %d", bn)
@@ -415,8 +415,8 @@ func (ob *Observer) ObserveTSSReceive(ctx context.Context, startBlock, toBlock u
 	return toBlock, nil
 }
 
-// CheckAndVoteInboundTokenZeta checks and votes on the given inbound Zeta token
-func (ob *Observer) CheckAndVoteInboundTokenZeta(
+// checkAndVoteInboundTokenZeta checks and votes on the given inbound Zeta token
+func (ob *Observer) checkAndVoteInboundTokenZeta(
 	ctx context.Context,
 	tx *client.Transaction,
 	receipt *ethtypes.Receipt,
@@ -437,7 +437,7 @@ func (ob *Observer) CheckAndVoteInboundTokenZeta(
 	}
 
 	// get zeta connector contract
-	addrConnector, connector, err := ob.GetConnectorContract()
+	addrConnector, connector, err := ob.getConnectorContract()
 	if err != nil {
 		return "", err
 	}
@@ -450,7 +450,7 @@ func (ob *Observer) CheckAndVoteInboundTokenZeta(
 			// sanity check tx event
 			err = common.ValidateEvmTxLog(&event.Raw, addrConnector, tx.Hash, common.TopicsZetaSent)
 			if err == nil {
-				msg = ob.BuildInboundVoteMsgForZetaSentEvent(app, event)
+				msg = ob.buildInboundVoteMsgForZetaSentEvent(app, event)
 			} else {
 				ob.Logger().Inbound.Error().Err(err).Msgf("CheckEvmTxLog error on inbound %s chain %d", tx.Hash, ob.Chain().ChainId)
 				return "", err
@@ -470,8 +470,8 @@ func (ob *Observer) CheckAndVoteInboundTokenZeta(
 	return msg.Digest(), nil
 }
 
-// CheckAndVoteInboundTokenERC20 checks and votes on the given inbound ERC20 token
-func (ob *Observer) CheckAndVoteInboundTokenERC20(
+// checkAndVoteInboundTokenERC20 checks and votes on the given inbound ERC20 token
+func (ob *Observer) checkAndVoteInboundTokenERC20(
 	ctx context.Context,
 	tx *client.Transaction,
 	receipt *ethtypes.Receipt,
@@ -487,7 +487,7 @@ func (ob *Observer) CheckAndVoteInboundTokenERC20(
 	}
 
 	// get erc20 custody contract
-	addrCustody, custody, err := ob.GetERC20CustodyContract()
+	addrCustody, custody, err := ob.getERC20CustodyContract()
 	if err != nil {
 		return "", err
 	}
@@ -501,7 +501,7 @@ func (ob *Observer) CheckAndVoteInboundTokenERC20(
 			// sanity check tx event
 			err = common.ValidateEvmTxLog(&zetaDeposited.Raw, addrCustody, tx.Hash, common.TopicsDeposited)
 			if err == nil {
-				msg = ob.BuildInboundVoteMsgForDepositedEvent(zetaDeposited, sender)
+				msg = ob.buildInboundVoteMsgForDepositedEvent(zetaDeposited, sender)
 			} else {
 				ob.Logger().Inbound.Error().Err(err).Msgf("CheckEvmTxLog error on inbound %s chain %d", tx.Hash, ob.Chain().ChainId)
 				return "", err
@@ -521,8 +521,8 @@ func (ob *Observer) CheckAndVoteInboundTokenERC20(
 	return msg.Digest(), nil
 }
 
-// CheckAndVoteInboundTokenGas checks and votes on the given inbound gas token
-func (ob *Observer) CheckAndVoteInboundTokenGas(
+// checkAndVoteInboundTokenGas checks and votes on the given inbound gas token
+func (ob *Observer) checkAndVoteInboundTokenGas(
 	ctx context.Context,
 	tx *client.Transaction,
 	receipt *ethtypes.Receipt,
@@ -547,7 +547,7 @@ func (ob *Observer) CheckAndVoteInboundTokenGas(
 	sender := ethcommon.HexToAddress(tx.From)
 
 	// build inbound vote message and post vote
-	msg := ob.BuildInboundVoteMsgForTokenSentToTSS(tx, sender, receipt.BlockNumber.Uint64())
+	msg := ob.buildInboundVoteMsgForTokenSentToTSS(tx, sender, receipt.BlockNumber.Uint64())
 	if msg == nil {
 		// donation, restricted tx, etc.
 		ob.Logger().Inbound.Info().Msgf("no vote message built for inbound %s chain %d", tx.Hash, ob.Chain().ChainId)
@@ -560,8 +560,8 @@ func (ob *Observer) CheckAndVoteInboundTokenGas(
 	return msg.Digest(), nil
 }
 
-// BuildInboundVoteMsgForDepositedEvent builds a inbound vote message for a Deposited event
-func (ob *Observer) BuildInboundVoteMsgForDepositedEvent(
+// buildInboundVoteMsgForDepositedEvent builds a inbound vote message for a Deposited event
+func (ob *Observer) buildInboundVoteMsgForDepositedEvent(
 	event *erc20custody.ERC20CustodyDeposited,
 	sender ethcommon.Address,
 ) *types.MsgVoteInbound {
@@ -616,8 +616,8 @@ func (ob *Observer) BuildInboundVoteMsgForDepositedEvent(
 	)
 }
 
-// BuildInboundVoteMsgForZetaSentEvent builds a inbound vote message for a ZetaSent event
-func (ob *Observer) BuildInboundVoteMsgForZetaSentEvent(
+// buildInboundVoteMsgForZetaSentEvent builds a inbound vote message for a ZetaSent event
+func (ob *Observer) buildInboundVoteMsgForZetaSentEvent(
 	appContext *zctx.AppContext,
 	event *zetaconnector.ZetaConnectorNonEthZetaSent,
 ) *types.MsgVoteInbound {
@@ -669,8 +669,8 @@ func (ob *Observer) BuildInboundVoteMsgForZetaSentEvent(
 	)
 }
 
-// BuildInboundVoteMsgForTokenSentToTSS builds a inbound vote message for a token sent to TSS
-func (ob *Observer) BuildInboundVoteMsgForTokenSentToTSS(
+// buildInboundVoteMsgForTokenSentToTSS builds a inbound vote message for a token sent to TSS
+func (ob *Observer) buildInboundVoteMsgForTokenSentToTSS(
 	tx *client.Transaction,
 	sender ethcommon.Address,
 	blockNumber uint64,
@@ -719,8 +719,8 @@ func (ob *Observer) BuildInboundVoteMsgForTokenSentToTSS(
 	)
 }
 
-// ObserveTSSReceiveInBlock queries the incoming gas asset to TSS address in a single block and posts votes
-func (ob *Observer) ObserveTSSReceiveInBlock(ctx context.Context, blockNumber uint64) error {
+// observeTSSReceiveInBlock queries the incoming gas asset to TSS address in a single block and posts votes
+func (ob *Observer) observeTSSReceiveInBlock(ctx context.Context, blockNumber uint64) error {
 	block, err := ob.GetBlockByNumberCached(ctx, blockNumber)
 	if err != nil {
 		return errors.Wrapf(err, "error getting block %d for chain %d", blockNumber, ob.Chain().ChainId)
@@ -733,7 +733,7 @@ func (ob *Observer) ObserveTSSReceiveInBlock(ctx context.Context, blockNumber ui
 				return errors.Wrapf(err, "error getting receipt for inbound %s chain %d", tx.Hash, ob.Chain().ChainId)
 			}
 
-			_, err = ob.CheckAndVoteInboundTokenGas(ctx, &tx, receipt, true)
+			_, err = ob.checkAndVoteInboundTokenGas(ctx, &tx, receipt, true)
 			if err != nil {
 				return errors.Wrapf(
 					err,