+++ zetaclient/chains/sui/observer/observer.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"os"
 	"strconv"
+	"sync"
 	"time"
 
 	"github.com/block-vision/sui-go-sdk/models"
@@ -19,6 +20,13 @@ type Observer struct {
 	*base.Observer
 	client  RPC
 	gateway *sui.Gateway
+
+	// nonce -> sui outbound tx
+	txMap map[uint64]models.SuiTransactionBlockResponse
+	txMu  sync.RWMutex
+
+	latestGasPrice uint64
+	gasPriceMu     sync.RWMutex
 }
 
 // RPC represents subset of Sui RPC methods.
@@ -41,8 +49,12 @@ func New(baseObserver *base.Observer, client RPC, gateway *sui.Gateway) *Observe
 		Observer: baseObserver,
 		client:   client,
 		gateway:  gateway,
+		txMap:    make(map[uint64]models.SuiTransactionBlockResponse),
 	}
 }
+
+// Gateway returns Sui gateway.
+func (ob *Observer) Gateway() *sui.Gateway { return ob.gateway }
 
 // CheckRPCStatus checks the RPC status of the chain.
 func (ob *Observer) CheckRPCStatus(ctx context.Context) error {
@@ -90,6 +102,8 @@ func (ob *Observer) PostGasPrice(ctx context.Context) error {
 	// no priority fee for Sui
 	const priorityFee = 0
 
+	ob.setLatestGasPrice(gasPrice)
+
 	_, err = ob.ZetacoreClient().PostVoteGasPrice(ctx, ob.Chain(), gasPrice, priorityFee, epochNum)
 	if err != nil {
 		return errors.Wrap(err, "unable to post vote for gas price")
@@ -98,10 +112,23 @@ func (ob *Observer) PostGasPrice(ctx context.Context) error {
 	return nil
 }
 
+func (ob *Observer) getLatestGasPrice() uint64 {
+	ob.gasPriceMu.RLock()
+	defer ob.gasPriceMu.RUnlock()
+
+	return ob.latestGasPrice
+}
+
+func (ob *Observer) setLatestGasPrice(price uint64) {
+	ob.gasPriceMu.Lock()
+	defer ob.gasPriceMu.Unlock()
+	ob.latestGasPrice = price
+}
+
 // ensureCursor ensures tx scroll cursor for inbound observations
-func (ob *Observer) ensureCursor() error {
+func (ob *Observer) ensureCursor() {
 	if ob.LastTxScanned() != "" {
-		return nil
+		return
 	}
 
 	// Note that this would only work for the empty chain database
@@ -109,36 +136,6 @@ func (ob *Observer) ensureCursor() error {
 	if envValue != "" {
 		ob.WithLastTxScanned(envValue)
 	}
-
-	return nil
-
-	//// let's take the first tx that was ever registered for the Gateway (deployment tx)
-	//// Note that this might have for a non-archival node
-	//req := models.SuiGetObjectRequest{
-	//	ObjectId: ob.gateway.PackageID(),
-	//	Options: models.SuiObjectDataOptions{
-	//		ShowPreviousTransaction: true,
-	//	},
-	//}
-	//
-	//res, err := ob.client.SuiGetObject(ctx, req)
-	//switch {
-	//case err != nil:
-	//	return errors.Wrap(err, "unable to get object")
-	//case res.Error != nil:
-	//	return errors.Errorf("get object error: %s (code %s)", res.Error.Error, res.Error.Code)
-	//case res.Data == nil:
-	//	return errors.New("object data is empty")
-	//case res.Data.PreviousTransaction == "":
-	//	return errors.New("previous transaction is empty")
-	//}
-	//
-	//cursor := client.EncodeCursor(models.EventId{
-	//	TxDigest: res.Data.PreviousTransaction,
-	//	EventSeq: "0",
-	//})
-	//
-	//return ob.setCursor(cursor)
 }
 
 func (ob *Observer) getCursor() string { return ob.LastTxScanned() }